+++ CHANGELOG.rst
@@ -4,6 +4,29 @@
 What's New
 **********
 
+
+==================================
+Version 1.5.0
+==================================
+
+----------------------------------
+Fixes
+----------------------------------
+- Fix version constants
+
+Additions
+----------------------------------
+- Added maya-2025 support
+- Added a workaround for additional . in the result of using listAttr with blendshape in Maya 2025.
+- Added support for PySide6 and shiboken6
+- Bumped PyQt from PyQt4 to PyQt6 and PyQt5
+
+----------------------------------
+Non-Backward Compatible Changes
+----------------------------------
+- Dropped support for maya-2020
+- Dropped support for PySide and shiboken
+
 ==================================
 Version 1.4.0
 ==================================